+++ src/main/kotlin/dev/furq/holodisplays/handlers/PacketHandler.kt
@@ -51,7 +51,7 @@ object PacketHandler {
         entityIds.getOrPut(player.uuid) { mutableMapOf() }
             .getOrPut(name) { mutableMapOf() }[displayRef] = entityId
 
-        val offsetPosition = position.add(display.displayType.offset.run { Vec3d(x, y, z) })
+        val offsetPosition = position.add(display.displayType.offset.run { Vec3d(x.toDouble(), y.toDouble(), z.toDouble()) })
 
         player.networkHandler.run {
             sendPacket(createSpawnPacket(entityId, offsetPosition, display.displayType, hologram))
